refactored composables

+++ composables/deletePuzzleSet.ts
@@ -1,13 +1,16 @@
 const key = "delete-puzzle-set";
 
-export const useDeletePuzzleSet = async (puzzleSet: any) => {
+export const useDeletePuzzleSet = () => {
   const { profile } = useFetchProfile();
 
-  return $fetch(
-    `/api/profiles/${profile.value!.id}/puzzle-sets/${puzzleSet.slug}`,
-    {
+  const deletePuzzleSet = async (puzzleSet: any) => {
+    if (!profile.value) throw createError("Profile is not loaded");
+
+    $fetch(`/api/profiles/${profile.value!.id}/puzzle-sets/${puzzleSet.slug}`, {
       method: "DELETE",
       key,
-    }
-  );
+    });
+  };
+
+  return { deletePuzzleSet };
 };
